Move component DataViewCustomInfoPanel from 'components' to 'components/index/CardsReference/PositiveNumberByDevelopedDate'
